Add files via upload
this is  a project i created on my own with the little knowledge i learned on python foundations past 2 month. i can say, i am proud of myself and am willing to learn more

check out the project, it's about guessing number game. i have made more good but am working on that, there are little things like when a player guesses the exact number, the loop has to start a fresh and let other player guess the number too, at the end of the day, player who has multiply guesses is considered the winner.
am working on that particular problem, please, feel free the offer you help, am willing to accept that.